Set UTF-8 encoding explicitly when loading a run file
I found that the sample_runs.json file was failing to load correctly due
to Unicode decoding errors. This might be a system-specific thing since I
think Python tries to pick a default encoding based on your OS environment
and I only noticed it on a Windows system so far.

sample_runs.json is UTF-8 and I'm guessing any other inputs will be too,
so this just forces that encoding.